Create new components, fix UI main screen
